tests: move tests to Functional folder
