+++ tests/Functional/UserTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Tests;
+namespace App\Tests\Functional;
 
 use ApiPlatform\Core\Bridge\Symfony\Bundle\Test\ApiTestCase;
 use Symfony\Contracts\HttpClient\Exception\TransportExceptionInterface;
